next-app(refactor): move fetcher function to utils/fetcher
