+++ packages/next-app/src/lib/ipfs.ts
@@ -11,18 +11,3 @@ export async function pinJSON(poolId: string, metadata: PoolMetadataAttribute[])
   });
   return await resp.json();
 }
-
-
-export async function fetchIpfsData(metadataCID: string) {
-  const response = await fetch(`https://ipfs.io/ipfs/${metadataCID}`);
-  return await response.json();
-}
-
-export async function fetcher<JSON = unknown>(
-  input: RequestInfo,
-  init?: RequestInit
-): Promise<JSON> {
-  const res = await fetch(input, init);
-  return res.json();
-}
-
