+++ packages/next-app/src/utils/fetcher.ts
@@ -1,0 +1,7 @@
+export async function fetcher<JSON = unknown>(
+  input: RequestInfo,
+  init?: RequestInit
+): Promise<JSON> {
+  const res = await fetch(input, init);
+  return res.json();
+}
